GPT-43 - Active Layers - Core mod for Active L Panel
In combination with work done on geoscience-portal project, the core has
been modified to draw the Active Layers panel up on the top-right of the
map page (with a collapse widget).  I rolled my own internal event
handling mechanism to defer the drawing of the OpenLayers Base Layers
(eg. "Google Satellite", ..) until the map
has been created.  I'd also started the function to just display the
OpenLayers Base Layer in its default location (on the map) so that other portal
clients don't have to implement this functionality, however it wasn't working and
is something I need to get back to.