GPT-138 handle spaces in query parameters
The portal-core code that constructs the URL already encodes the parameters for us, but it turns spaces into plus signs. That's actually correct but some servers reject them (they expect "%20" for spaces).
This change decodes the query segment of the URI and then uses the URI constructor which also encodes, but turns spaces into "%20".